+++ CHANGELOG.md
@@ -9,6 +9,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### TODO
 
-- Add linter to check for `NewXXX` to be after struct declaration but before struct methods.
-- Add linter to check for `MustXXX` to be after struct declaration but before struct methods.
-- Add linter to check that public methods are before private ones.
+- Added linter to check for `NewXXX` to be after struct declaration but before struct methods.
+- Added linter to check for `MustXXX` to be after struct declaration but before struct methods.
+- Add linter to check that exported (public) methods are before not exported (private) ones.